move xarray engine into driver file
