+++ forest_lite/server/drivers/xarray_h5netcdf.py
@@ -5,10 +5,13 @@ from forest_lite.server.drivers.base import BaseDriver
 from forest_lite.server.lib import core
 from pydantic import BaseModel, validator
 from typing import List
+from functools import lru_cache
+import datetime as dt
 
 
 class Settings(BaseModel):
     pattern: str
+    engine: str = "h5netcdf"
     data_vars: List[str] = None
 
 
@@ -23,7 +26,8 @@ driver = BaseDriver()
 @driver.override("data_tile")
 def data_tile(settings, data_var, z, x, y, query=None):
     pattern = Settings(**settings).pattern
-    return core.get_data_tile(pattern, data_var, z, x, y, query)
+    engine = Settings(**settings).engine
+    return get_data_tile(pattern, engine, data_var, z, x, y, query)
 
 
 @driver.override("description")
@@ -32,7 +36,7 @@ def description(settings):
     paths = sorted(glob.glob(settings.pattern))
     if len(paths) > 0:
         path = paths[-1]
-        with xarray.open_dataset(path, engine="h5netcdf") as nc:
+        with xarray.open_dataset(path, engine=settings.engine) as nc:
             data = nc.to_dict(data=False)
         # Filter data_vars
         if settings.data_vars is not None:
@@ -55,7 +59,7 @@ def points(settings, data_var, dim_name, query=None):
     data = []
     if len(paths) > 0:
         path = paths[-1]
-        with xarray.open_dataset(path, engine="h5netcdf") as nc:
+        with xarray.open_dataset(path, engine=settings.engine) as nc:
             attrs = nc[dim_name].attrs
             data = nc[dim_name].values
     return {
@@ -64,3 +68,66 @@ def points(settings, data_var, dim_name, query=None):
         "data": data,
         "attrs": attrs
     }
+
+
+def get_data_tile(pattern, engine, data_var, z, x, y, query=None):
+    if query is not None:
+        # Hashable query
+        query = frozenset(query.items())
+    path = core.get_path(pattern)
+    return _data_tile(path, engine, data_var, z, x, y, query)
+
+
+@lru_cache
+def _data_tile(path, engine, data_var, z, x, y, query):
+    zxy = (z, x, y)
+    with xarray.open_dataset(path, engine=engine) as nc:
+
+        # Find lons/lats related to data_var
+        var = nc[data_var]
+        for key in var.dims:
+            if key.startswith("longitude"):
+                lons = var[key].values
+            if key.startswith("latitude"):
+                lats = var[key].values
+
+        # Find 2D values array
+        if query is None:
+            array = nc[data_var]
+        else:
+            idx = dict(query)
+
+            # Map miliseconds to datetime
+            idx = { key: convert_ms(key, value) for key, value in idx.items() }
+
+            try:
+                array = nc[data_var].sel(**idx, method="nearest")
+            except ValueError:
+                idx = { key: int(value) for key, value in idx.items() }
+                array = nc[data_var].isel(**idx)
+
+        units = nc[data_var].units
+
+    assert array.ndim == 2, f"dims: {var.dims}"
+    values = array.values
+
+    # Mask moisture_content_of_soil_layer (TODO: Generalise)
+    if "moisture_content" in data_var:
+        fill_value = values.max()
+        values = np.ma.masked_equal(values, fill_value)
+    return _tile({
+        "longitude": lons,
+        "latitude": lats,
+        "values": values,
+        "units": units
+    }, z, x, y)
+
+
+def convert_ms(dim, value):
+    if "time" in dim.lower():
+        if isinstance(value, str):
+            return value
+        else:
+            return dt.datetime.fromtimestamp(value / 1000)
+    else:
+        return value
